+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Changes
 
 * Remove chromedriver `110.0.5481.77`
+* Try to start Chrome several times
 
 ## 1.17.0 (2023-04-13)